fix: search topic based on the query
